Move array internals to ember-metal
